App works with file move. Testing is horrible
